MINOR Conditional call for deleteTreeNodeJS() method thats been removed from trunk
git-svn-id: http://svn.silverstripe.com/open/modules/linkchecker/trunk@115127 467b73ca-7a2a-4603-9d3b-597d59a354a9

+++ code/LinkCheckAdmin.php
@@ -187,7 +187,8 @@ class LinkCheckAdmin extends LeftAndMain {
 			foreach($ids as $id) {
 				if(is_numeric($id)) {
 					$record = DataObject::get_by_id('LinkCheckRun', $id);
-					$script .= $this->deleteTreeNodeJS($record);
+					// Compatibility with 3.0
+					if($this->hasMethod('deleteTreeNodeJS')) $script .= $this->deleteTreeNodeJS($record);
 					$record->delete();
 					$record->destroy();
 				}
